requirements.txt
Flask==3.1.1
pymongo==4.13.2
python-dotenv==1.0.0
